refactor: introduce SerializedEventListener, fix delta types and editor tooling.
- Add SerializedEventListener as a [GlobalClass, Tool] specialization of
  SerializedReference<IEventListener> for pipeline serialization
- Add NewableSpawner<T> for types with parameterless constructors
- Fix GodotLoop event delta type from double to float to match Godot
  conventions (casts via explicit (float) conversion)
- Replace SerializedReference<IContext> with SerializedContext in
  entity adapters and game pipeline
- Improve SerializedInstallInfo editor UX: auto-detect NewableSpawner
  when type has empty ctor, support int-based binding selection,
  extract type-resolution helpers to class-level methods
- Add AssemblyCache.LoadPlugins() to discover and load addon DLLs
  via AssemblyLoadContext in editor builds
- Add [Tool] attribute to NodeInstaller and ResourceInstaller
- Fix inverted null checks in HasInjectConstructor and ternary
  fallback conditions in Spawner/Controller properties
- Add null-conditional disposal and lazy context initialization
- Add GD.PushWarning for unconfigured event listeners in pipelines
- Add GUBBINS_ENABLED guard to TypeKind and clean up unused usings

+++ Gubbins.Godot/Runtime/Events/Builtin/LoopEvents.cs
@@ -45,10 +45,10 @@ public static class LoopEvents
 /// <summary>
 /// Base class for WindowLoop event wrappers.
 /// </summary>
-public class LoopEvent : IEventSubscriable<double>
+public class LoopEvent : IEventSubscriable<float>
 {
     private readonly GodotLoop.Kind                                    m_Kind;
-    private readonly Dictionary<IEventHandler<double>, Action<double>> m_HandlerMap = new();
+    private readonly Dictionary<IEventHandler<float>, Action<float>> m_HandlerMap = new();
 
     internal LoopEvent(GodotLoop.Kind kind)
     {
@@ -56,10 +56,10 @@ public class LoopEvent : IEventSubscriable<double>
     }
 
     /// <inheritdoc/>
-    public void Subscribe(IEventHandler<double> handler)
+    public void Subscribe(IEventHandler<float> handler)
     {
-        Action<double> action;
-        if (handler is ActionEventHandler<double> actionHandler)
+        Action<float> action;
+        if (handler is ActionEventHandler<float> actionHandler)
         {
             action = actionHandler.Invocation;
         }
@@ -73,7 +73,7 @@ public class LoopEvent : IEventSubscriable<double>
     }
 
     /// <inheritdoc/>
-    public bool Unsubscribe(IEventHandler<double> handler)
+    public bool Unsubscribe(IEventHandler<float> handler)
     {
         if (!m_HandlerMap.Remove(handler, out var action)) return false;
         GodotLoop.UnregisterUpdate(m_Kind, action);
